lib: refactor wc_cnx_t => wc_context_t and user data
Note: this is an itermediary step in a relatively big SDK refactoring, the
documentation is not up to date and the code might even not work.

We don't expose the concept of connection anymore, but rather the
concept of context. Indeed, we don't want to tie all the metadata beared
by this object to a given effective conenction (file descriptor), as the
datasync service users probably want to cope well with connection loss,
and temporarilly offline situations.

While this makes the SDK more usable in real world conditions, it adds more
work to the SDK user to deal with the file descriptors to poll: the SDK
can ask to the application to start or stop polling a file descriptor at
any time or change the type of events (new events:
WC_EVENT_{ADD,DEL,MODIFY}_FD). This tends to make the "boilerplate" code to
incorporate the SDK in your own event loop much bigger. To address this, we
will offer in a very near future the ability to integrate transparently
with your existing libevent, libev or libuv event loop, stay tuned...

Smaller change: the user data associated to a context may now be obtained
via:

    wc_context_get_user_data(ctx);