lib: integrate with some popular event loops
This patch adds the ability to defer all the dirty work of catching and
dealing with the SDKs WC_EVENT_*_FD events to one of three popular event
loop library:

- libev
- libuv
- libevent

The SDK will only require a reference to the event loop and a couple of
callbacks, and will make your life easier than before, as you don't need to
deal with any file descriptor yourself, and it will also take care of
sendng the keepalive packet on a regular basis.

If your favorite event loop is not one of the three above, it's fairly
easy to add a new integration, so you can open an issue or a pull request,
you can also use the WC_EVENT_*_FD events in your program.

To see an example of how to integrate with libev, look at the updated
legorange example.