lib: handle both readable and writable events
In order to use lws correctly, in particular regarding how to close a
client websockets, we need to use lws_callback_on_writable() and then
lws_service_fd() with the correct informations so that lws will send the
LWS_CALLBACK_CLIENT_WRITEABLE event, to which we reply -1.

We got rid of the too restrictive wc_cnx_on_readable() API function, and
now wc_handle_fd_events() must be called.